requirements.txt
# incomplete
flask
flask-sqlalchemy
flask-migrate
flask-login
flask-mail
pyjwt
python-dotenv
